+++ web/app/themes/helsinki-sovittelutoiminta/resources/views/blocks/question-and-answer.blade.php
@@ -1,0 +1,8 @@
+<div class="{{ esc_attr($block->classes) }}">
+  <div class="wp-block-hds-question-and-answer__question">
+    {{ $question }}
+  </div>
+  <div class="wp-block-hds-question-and-answer__answer">
+    {!! $content !!}
+  </div>
+</div>